requirements.txt
# Requirements for dataset tokenization and training
tiktoken>=0.5.0
torch>=2.0.0
numpy>=1.24.0

# Web API Server
flask>=2.3.0
flask-cors>=4.0.0

# Multi-modal AI - Screenshot Analysis
opencv-python>=4.8.1
pillow>=10.1.0
pytesseract>=0.3.10

# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.0
